backend/requirements/base.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-filter>=24.3,<25.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<12.0
google-api-python-client>=2.149,<3.0
google-auth>=2.35,<3.0
google-auth-httplib2>=0.2,<1.0
